Token lifetime serialization uses doubles.

+++ NET-Core/LibUA/Client.cs
@@ -206,7 +206,7 @@ namespace LibUA
                 succeeded &= sendBuf.Encode(securityTokenRequestType);
                 succeeded &= sendBuf.Encode(messageSecurityMode);
                 succeeded &= sendBuf.EncodeUAByteString(clientNonce);
-                succeeded &= sendBuf.Encode(reqLifetime);
+                succeeded &= sendBuf.Encode((double)reqLifetime);
 
                 config.LocalNonce = clientNonce;
 
@@ -382,7 +382,7 @@ namespace LibUA
                 if (!recvHandler.RecvBuf.Decode(out uint channelId)) { return StatusCode.BadDecodingError; }
                 if (!recvHandler.RecvBuf.Decode(out uint tokenId)) { return StatusCode.BadDecodingError; }
                 if (!recvHandler.RecvBuf.Decode(out ulong createAtTimestamp)) { return StatusCode.BadDecodingError; }
-                if (!recvHandler.RecvBuf.Decode(out uint respLifetime)) { return StatusCode.BadDecodingError; }
+                if (!recvHandler.RecvBuf.Decode(out double respLifetime)) { return StatusCode.BadDecodingError; }
                 if (!recvHandler.RecvBuf.DecodeUAByteString(out byte[] serverNonce)) { return StatusCode.BadDecodingError; }
 
                 if (renew)
@@ -396,7 +396,7 @@ namespace LibUA
                 config.TokenCreatedAt = DateTimeOffset.FromFileTime((long)createAtTimestamp);
                 if (config.TokenLifetime == 0)
                 {
-                    config.TokenLifetime = respLifetime;
+                    config.TokenLifetime = (uint)respLifetime;
                 }
                 config.RemoteNonce = serverNonce;
